Don't override optimization setting in debug builds.
The cmake configuration for this package unconditionally sets the optimization
flag to -O3.  It does this, however, even for DEBUG builds, where that
is probably not the desired behavior (it took me by surprise, anyway).

Change to override the optimization to -O3 for RELEASE and RELWITHDEBINFO
builds, but not for DEBUG.